[feat] : API 분리
- 파일 컨벤션에 맞게 수정
- API 분리
- 불필요한 파일 삭제
