+++ src/components/TB/OrgModal.js
@@ -1,9 +1,10 @@
+import { useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
+
 import styled from 'styled-components';
 
 import {AiOutlineSearch, AiFillFolderOpen} from 'react-icons/ai';
 import {LuBuilding2} from 'react-icons/lu';
-import { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
 
 import {orgTreeApi, orgEmpListApi, searchOrg} from '../../utils/API';
 
@@ -13,7 +14,6 @@ export default function OrgModal(props){
   
   const [data, setData] = useState(JSON.parse('[{"id":"", "name":""}]'));
   const [empList, setEmpList] = useState(JSON.parse('[{"id":"", "name":""}]'));
-  const [open, setOpen] = useState(Array(data.length).fill(false));
   const emp_id = useSelector(state => state.gnbMenu.key);
 
   const [searchOption, setSearchOption] = useState("");
@@ -21,11 +21,9 @@ export default function OrgModal(props){
   const [detail, setDetail] = useState();
   const [detailOpen, setDetailOpen] = useState();
 
-  console.log('in orgModal', open)
   useEffect(() => {
     async function LoadData(emp_id){
       const res = await orgTreeApi('comp', "","",emp_id);
-      console.log('in useEffect : ', res.data.data)
       setData(res.data.data);
     }
     LoadData(emp_id);
@@ -37,149 +35,132 @@ export default function OrgModal(props){
   }
 
   function EmpDetailHandler(detailData){
-    console.log('in handler : ', detailData);
     setDetail(detailData);
     setDetailOpen(!detailOpen);
   }
 
   function searchHandler(){
-    
     searchOrg(searchOption, searchText).then(res => {
-      console.log(res.data.data);
       if(searchOption === 'dept'){
-        console.log('type dept');
         setData(res.data.data);
       }
       if(searchOption === 'emp'){
-        console.log('type emp');
         setEmpList(res.data.data);
       }
     });
   }
 
-    return (
-      <ModalBackdrop onClick={() => {props.api('org')}}>
-      <OrgModalView onClick={(e) => e.stopPropagation()}>
-        <div id="nav">
-          <h3>조직도</h3>
-          <div onClick={() => {props.api('org')}}>X</div>
-        </div>
-        <hr style={{width: '965px'}}/>
-        <div>
-          <div id='search'>
-            <div>
-              <select value={searchOption} onChange={(e)=>setSearchOption(e.target.value)} ><option value='all'>전체</option><option value='dept'>부서명</option><option value='emp'>사원명</option></select>
-              <input type="text" placeholder="검색" value={searchText} onChange={(e)=>setSearchText(e.target.value)}/>
-            </div>
-            <AiOutlineSearch onClick={() => {searchHandler()}}/>
+  return (
+    <ModalBackdrop onClick={() => {props.api('org')}}>
+    <OrgModalView onClick={(e) => e.stopPropagation()}>
+      <div id="nav">
+        <h3>조직도</h3>
+        <div onClick={() => {props.api('org')}}>X</div>
+      </div>
+      <hr style={{width: '965px'}}/>
+      <div>
+        <div id='search'>
+          <div>
+            <select value={searchOption} onChange={(e)=>setSearchOption(e.target.value)} ><option value='all'>전체</option><option value='dept'>부서명</option><option value='emp'>사원명</option></select>
+            <input type="text" placeholder="검색" value={searchText} onChange={(e)=>setSearchText(e.target.value)}/>
           </div>
-          <div id='content' className="flex">
-            <div id='deptList'>
-            { false &&
-              data.map((a, i) => (
-                <div key={a['name']+a['id']}>
-                  <div onClick={(i) => {}}><LuBuilding2 />{a['name']}</div>
-                  <DeptList value={a['id']} api={loadEmpList}/>
-                </div>
-                ))
-              }
-              
-            </div>
-            
-              <EmpList value={empList} api={EmpDetailHandler}/>
-            
-            <div id='empDetail'>
-              {detailOpen && <EmpDetail list={detail}/>}
+          <AiOutlineSearch onClick={() => {searchHandler()}}/>
+        </div>
+        <div id='content' className="flex">
+          <div id='deptList'>
+          { false &&
+            data.map((a, i) => (
+              <div key={a['name']+a['id']}>
+                <div onClick={(i) => {}}><LuBuilding2 />{a['name']}</div>
+                <DeptList value={a['id']} api={loadEmpList}/>
+              </div>
+              ))
+            }
           </div>
+            <EmpList value={empList} api={EmpDetailHandler}/>
+          <div id='empDetail'>
+            {detailOpen && <EmpDetail list={detail}/>}
         </div>
-        </div>
-      </OrgModalView>
-    </ModalBackdrop>
-  
-    )
-  }
+      </div>
+      </div>
+    </OrgModalView>
+  </ModalBackdrop>
+  );
+}
 
+export function DeptList(props) {
+  const [data, setData] = useState(['[{"id":"", "name":""}]']);
+  const [deptOpen, setDeptOpen] = useState([false]);
 
-  export function DeptList(props) {
-    const [data, setData] = useState(['[{"id":"", "name":""}]']);
-    const [deptOpen, setDeptOpen] = useState([false]);
-    console.log('in deptList')
-  
-    useEffect(() => {
-      async function LoadData(){
-        const res = await orgTreeApi('Dept1', props.value,"","");
-        console.log('in useEffect : ', res.data.data)
-        setData(res.data.data);
-      }
-      LoadData();
-    }, [props.value]);
-  
-    return (
-      <>
-        {data.length > 0 && (
-          data.map((a, i) => (
-            <div key={a['name']+a['id']} >
-              <div onClick={() => {
-                setDeptOpen(deptOpen => {
-                  const newDeptOpen = [...deptOpen];
-                  newDeptOpen[i] = !newDeptOpen[i];
-                  return newDeptOpen;
-                });
-                props.api('dept', a['id']);
-                }} style={{marginLeft:'15px'}}><AiFillFolderOpen />{a['name']}
-              </div>
-              { deptOpen[i] && (
-                  <Dept dept={a['id']} comp={props.value} api={props.api}/>
-                )}
+  useEffect(() => {
+    async function LoadData(){
+      const res = await orgTreeApi('Dept1', props.value,"","");
+      setData(res.data.data);
+    }
+    LoadData();
+  }, [props.value]);
+
+  return (
+    <>
+      {data.length > 0 && (
+        data.map((a, i) => (
+          <div key={a['name']+a['id']} >
+            <div onClick={() => {
+              setDeptOpen(deptOpen => {
+                const newDeptOpen = [...deptOpen];
+                newDeptOpen[i] = !newDeptOpen[i];
+                return newDeptOpen;
+              });
+              props.api('dept', a['id']);
+              }} style={{marginLeft:'15px'}}><AiFillFolderOpen />{a['name']}
             </div>
-          )))
-        }
-      </>
-    )
-  }
-  
-    
-  export function Dept(props) {
-    const [data, setData] = useState(['[{"id":"", "name":""}]']);
-    const [dept, setDept] = useState([false]);
-    console.log('in deptList')
-  
-    useEffect(() => {
-      async function LoadData(){
-        const res = await orgTreeApi('Dept2', props.comp,props.dept,"");
-        console.log('in useEffect : ', res.data.data)
-        setData(res.data.data);
+            { deptOpen[i] && (
+                <Dept dept={a['id']} comp={props.value} api={props.api}/>
+              )}
+          </div>
+        )))
       }
-      LoadData();
-    }, [props]);
+    </>
+  );
+}
   
-    return (
-      <>
-        {data.length > 0 && (
-          data.map((a, i) => (
-            <div key={a['name']+a['id']} >
-              <div onClick={() => {
-                setDept(deptOpen => {
-                  const newDept = [...dept];
-                  newDept[i] = !newDept[i];
-                  return newDept;
-                });
-                props.api('dept', a['id']);
-                }} style={{marginLeft:'35px'}}><AiFillFolderOpen />{a['name']}
-              </div>
-              { dept[i] && (
-                  <Dept dept={a['id']} comp={props.comp}/>
-                )}
+export function Dept(props) {
+  const [data, setData] = useState(['[{"id":"", "name":""}]']);
+  const [dept, setDept] = useState([false]);
+
+  useEffect(() => {
+    async function LoadData(){
+      const res = await orgTreeApi('Dept2', props.comp,props.dept,"");
+      setData(res.data.data);
+    }
+    LoadData();
+  }, [props]);
+
+  return (
+    <>
+      {data.length > 0 && (
+        data.map((a, i) => (
+          <div key={a['name']+a['id']} >
+            <div onClick={() => {
+              setDept(deptOpen => {
+                const newDept = [...dept];
+                newDept[i] = !newDept[i];
+                return newDept;
+              });
+              props.api('dept', a['id']);
+              }} style={{marginLeft:'35px'}}><AiFillFolderOpen />{a['name']}
             </div>
-          )))
-        }
-      </>
-    )
-  }
-  
+            { dept[i] && (
+                <Dept dept={a['id']} comp={props.comp}/>
+              )}
+          </div>
+        )))
+      }
+    </>
+  );
+}
 
 export function EmpDetail(props) {
-  console.log('emp detail : ', props.list)
   return (
     <DetailEmp>
       <div>
@@ -198,7 +179,7 @@ export function EmpDetail(props) {
         </tr>
       </table>
     </DetailEmp>
-  )
+  );
 }
 
 export const ModalBackdrop = styled.div`
@@ -297,7 +278,6 @@ export const OrgModalView = styled.div`
   }
 `;
 export const DetailEmp = styled.div`
-
 > div {
   margin: 10px;
   text-align: center;
