+++ src/pages/TB.js
@@ -6,15 +6,69 @@ import { AiOutlineSearch, AiOutlineBell, AiOutlineMore, AiOutlineDeploymentUnit 
 
 import {profile, search, alert, org, set} from '../utils/Slice';
 
-import Profile from '../components/TB/Profile'
-
+import Profile from '../components/TB/Profile';
 import ProfileModal from '../components/TB/ProfileModal';
-import SearchModal from '../components/TB/SearchModal';
-import AlertModal from '../components/TB/AlertModal';
 import OrgModal from '../components/TB/OrgModal';
-import SetModal from '../components/TB/SetModal';
+import { SearchModal, AlertModal, SetModal } from '../components/TB/Modal';
 
-export const StyledTB = styled.div`
+
+export default function TB() {
+  const dispatch = useDispatch();
+  const profileOpen = useSelector(state => state.modalSwitch.profile);
+  const searchOpen = useSelector(state => state.modalSwitch.search);
+  const alertOpen = useSelector(state => state.modalSwitch.alert);
+  const orgOpen = useSelector(state => state.modalSwitch.org);
+  const setOpen = useSelector(state => state.modalSwitch.set);
+  
+  // modal을 키고 끄는 메서드
+  function ModalSwitch(type){
+    if(type === 'profile'){
+      dispatch(profile());
+    }
+    if(type === 'search'){
+      dispatch(search());
+    }
+    if(type === 'alert'){
+      dispatch(alert());
+    }
+    if(type === 'org'){
+      dispatch(org());
+    }
+    if(type === 'set'){
+      dispatch(set());
+    }
+  }
+  
+  return (
+    <TBArea>
+      <div id='logo'>
+        <Link to='/'>
+        <h1>Amaranth10</h1>
+        </Link>
+      </div>
+
+      <div>
+        <div id='prf' onClick={() => {ModalSwitch('profile');}}>        
+          <Profile />
+          {profileOpen && <ProfileModal api={ModalSwitch}/>}
+        </div>
+
+        <IconArea>
+          <AiOutlineSearch onClick={() => {ModalSwitch('search');}}/>
+          <AiOutlineBell onClick={() => {ModalSwitch('alert');}}/>
+          <AiOutlineDeploymentUnit onClick={() => {ModalSwitch('org');}}/>
+          <AiOutlineMore onClick={() => {ModalSwitch('set');}}/>
+        </IconArea>
+        {searchOpen && <SearchModal api={ModalSwitch} />}
+        {alertOpen && <AlertModal api={ModalSwitch}/>}
+        {orgOpen && <OrgModal api={ModalSwitch}/>}
+        {setOpen && <SetModal api={ModalSwitch}/>}
+      </div>
+    </TBArea>
+  );
+}
+
+export const TBArea = styled.div`
 display: flex;
 justify-content: space-between;
 position: relative;
@@ -46,8 +100,8 @@ width:100%;
   }
  
 }
-`
-export const IconDiv = styled.div`
+`;
+export const IconArea = styled.div`
 padding-top: 30px;
 width:180px;
 position: relative;
@@ -57,60 +111,4 @@ right:50px;
   height:30px;
   margin:5px;
 }
-`
-
-export default function TB() {
-  const dispatch = useDispatch();
-  const profileOpen = useSelector(state => state.modalSwitch.profile);
-  const searchOpen = useSelector(state => state.modalSwitch.search);
-  const alertOpen = useSelector(state => state.modalSwitch.alert);
-  const orgOpen = useSelector(state => state.modalSwitch.org);
-  const setOpen = useSelector(state => state.modalSwitch.set);
-  
-  // modal을 키고 끄는 메서드
-  function ModalSwitch(type){
-    if(type === 'profile'){
-      dispatch(profile());
-    }
-    if(type === 'search'){
-      dispatch(search());
-    }
-    if(type === 'alert'){
-      dispatch(alert());
-    }
-    if(type === 'org'){
-      dispatch(org());
-    }
-    if(type === 'set'){
-      dispatch(set());
-    }
-  }
-  
-  return (
-    <StyledTB>
-      <div id='logo'>
-        <Link to='/'>
-        <h1>Amaranth10</h1>
-        </Link>
-      </div>
-
-      <div>
-        <div id='prf' onClick={() => {ModalSwitch('profile')}}>        
-          <Profile />
-          {profileOpen && <ProfileModal api={ModalSwitch}/>}
-        </div>
-
-        <IconDiv >
-          <AiOutlineSearch onClick={() => {ModalSwitch('search')}}/>
-          <AiOutlineBell onClick={() => {ModalSwitch('alert')}}/>
-          <AiOutlineDeploymentUnit onClick={() => {ModalSwitch('org')}}/>
-          <AiOutlineMore onClick={() => {ModalSwitch('set')}}/>
-        </IconDiv>
-        {searchOpen && <SearchModal api={ModalSwitch} />}
-        {alertOpen && <AlertModal api={ModalSwitch}/>}
-        {orgOpen && <OrgModal api={ModalSwitch}/>}
-        {setOpen && <SetModal api={ModalSwitch}/>}
-      </div>
-    </StyledTB>
-  );
-}
+`;
