+++ src/pages/VIEW.js
@@ -1,38 +1,38 @@
 export function VIEW() {
-    return (
-      <div className="view">
-          <h1>VIEW</h1>
-      </div>
-    );
-  }
-  
- export function Test() {
-    return (
-      <div>
-          <h1>공지사항</h1>
-          
-      </div>
-    );
-  }
+  return (
+    <div className="view">
+        <h1>VIEW</h1>
+    </div>
+  );
+}
 
-  export function EmpM(){
-    return (
-      <>
-      <h1>사원 관리 페이지 입니다.</h1>
-      </>
-    )
-  }
-  export function DeptM(){
-    return (
-      <>
-      <h1>부서 관리 페이지 입니다.</h1>
-      </>
-    )
-  }
-  export function CompM(){
-    return (
-      <>
-      <h1>회사 관리 페이지 입니다.</h1>
-      </>
-    )
-  }
+export function Test() {
+  return (
+    <div>
+        <h1>공지사항</h1>
+        
+    </div>
+  );
+}
+
+export function EmpM(){
+  return (
+    <>
+    <h1>사원 관리 페이지 입니다.</h1>
+    </>
+  );
+}
+export function DeptM(){
+  return (
+    <>
+    <h1>부서 관리 페이지 입니다.</h1>
+    </>
+  );
+}
+export function CompM(){
+  return (
+    <>
+    <h1>회사 관리 페이지 입니다.</h1>
+    </>
+  );
+}
