+++ src/utils/API.js
@@ -28,8 +28,6 @@ export function GnbFavorDeleteApi(emp_id, menu_id){
             baseURL : 'http://localhost:8080'
         });
     } 
-
-    console.log('some param is null ..')
     return Promise.reject(new Error("Invalid parameters"));
 }
 
@@ -41,13 +39,11 @@ export function profileAPI(emp_id){
             baseURL : 'http://localhost:8080'
         });
     } else {
-        console.log('some param is null ..')
         return Promise.reject(new Error("Invalid parameters"));
     }
 }
         
 export function orgTreeApi(type = "", comp_id = "", dept_id = "", emp_id = ""){  
-    console.log("type : ", type," comp_id : ",  comp_id," dept_id : ",  dept_id," emp_id : ",  emp_id)
     return axios({
         url: `/Modal/org/tree?type=${type}&compId=${comp_id}&deptId=${dept_id}&empId=${emp_id}`,
         method:'get',
@@ -56,9 +52,7 @@ export function orgTreeApi(type = "", comp_id = "", dept_id = "", emp_id = ""){
 }
 
 export function orgEmpListApi(type = "", text = ""){  
-    console.log("type : ", type," text : ",  text)
     if (type === 'comp'){
-        console.log('type is comp')
         return axios({
             url: `/Modal/org/empList?type=${type}&compId=${text}`,
             method:'get',
@@ -66,7 +60,6 @@ export function orgEmpListApi(type = "", text = ""){
         });
     }
     if (type === 'dept'){
-        console.log('type is dept')
         return axios({
             url: `/Modal/org/empList?type=${type}&deptId=${text}`,
             method:'get',
@@ -75,9 +68,7 @@ export function orgEmpListApi(type = "", text = ""){
     }
 }
 
-
 export function menuDetailApi(dept_id){  
-    console.log("dept_id : ", dept_id)
     return axios({
         url: `/Modal/org/empList?deptId=${dept_id}`,
         method:'get',
@@ -87,7 +78,13 @@ export function menuDetailApi(dept_id){
 }
 
 export function searchOrg(type, text){  
-    console.log("dept_id : ", type, text)
+    if (type === 'all') {
+        return axios({
+            url: `/Modal/org/search?type=${type}&text=%25${text}%25`,
+            method:'get',
+            baseURL : 'http://localhost:8080'
+        });
+    }
     if (type === 'dept') {
         return axios({
             url: `/Modal/org/search?type=${type}&text=%25${text}%25`,
@@ -102,6 +99,142 @@ export function searchOrg(type, text){
             baseURL : 'http://localhost:8080'
         });
     }
-    
-    
+}
+
+export function GnbApi(){
+    return axios({
+      url: `/setting/menu/gnb`,
+      method:'get',
+      baseURL : 'http://localhost:8080',
+    })
+}
+
+export function FavorApi(type, empId, menuId){
+    if (type === 'load') {
+        return axios({
+        url: `/setting/favor?empId=${empId}&menuId=${menuId}`,
+        method:'get',
+        baseURL : 'http://localhost:8080',
+        })
+    }
+    if (type === 'off') {
+        return axios({
+            url: `/setting/favor?empId=${empId}&menuId=${menuId}`,
+            method:'delete',
+            baseURL : 'http://localhost:8080'
+        })
+    }
+    if (type === 'on') {
+        return axios({
+            url: `/setting/favor?empId=${empId}&menuId=${menuId}`,
+            method:'post',
+            baseURL : 'http://localhost:8080',
+            data : {
+              empId: empId,
+              menuId: menuId
+            }
+          })
+    }
+}
+
+export function searchAPI(formData) {
+    return axios({
+        url: `/setting/menu/search?gnbName=${formData.get("gnbName")}%25&name=%25${formData.get("name")}%25`,
+        method:'get',
+        baseURL : 'http://localhost:8080',
+        data: {
+          formData
+        }
+    });
+}
+
+export function saveMenuAPI(formData, data, type) {
+    if (type === '1'){
+        return axios({
+            url: `/setting/menu?type=1`,
+            method:'post',
+            baseURL : 'http://localhost:8080',
+            data:{
+              parId: data['parId'],
+              name: formData.get('name') === "" ? data['name'] : formData.get('name'),
+              enabledYN:  formData.get('enabledYN') === null ? data['enabledYN'] : formData.get('enabledYN'),
+              sortOrder:  formData.get('sortOrder') === "" ? data['sortOrder'] : formData.get('sortOrder'),
+              iconUrl:  formData.get('iconUrl') === "" ? data['iconUrl'] : formData.get('iconUrl')
+            }
+        });
+    }
+    if (type === '2'){
+        return axios({
+            url: `/setting/menu?type=2`,
+            method:'post',
+            baseURL : 'http://localhost:8080',
+            data:{
+              id:  data['id'],
+              parId: data['parId'],
+              name: formData.get('name') === "" ? data['name'] : formData.get('name'),
+              enabledYN:  formData.get('enabledYN') === null ? data['enabledYN'] : formData.get('enabledYN'),
+              sortOrder:  formData.get('sortOrder') === "" ? data['sortOrder'] : formData.get('sortOrder'),
+              iconUrl:  formData.get('iconUrl') === "" ? data['iconUrl'] : formData.get('iconUrl')
+            }
+        });
+    }
+    if (type === '3'){
+        return axios({
+            url: `/setting/menu?type=${type}`,
+            method:'post',
+            baseURL : 'http://localhost:8080',
+            data:{
+              parId: 72, // 상위메뉴 만들기 전까지 고정
+              name: formData.get('name') === "" ? data['name'] : formData.get('name'),
+              enabledYN:  formData.get('enabledYN') === null ? data['enabledYN'] : formData.get('enabledYN'),
+              sortOrder:  formData.get('sortOrder') === "" ? data['sortOrder'] : formData.get('sortOrder')
+            }
+        });
+    }
+    if (type === '4') {
+        return axios({
+            url: `/setting/menu?type=4`,
+            method:'post',
+            baseURL : 'http://localhost:8080',
+            data:{
+              id:  data['id'],
+              parId: formData.get('parId') === null ? data['parId'] : formData.get('parId'),
+              name: formData.get('name') === "" ? data['name'] : formData.get('name'),
+              enabledYN:  formData.get('enabledYN') === null ? data['enabledYN'] : formData.get('enabledYN'),
+              sortOrder:  formData.get('sortOrder') === "" ? data['sortOrder'] : formData.get('sortOrder')
+            }
+        });
+
+    }   
+}
+      /*
+      // 이미지 데이터 보내기
+      if (iconFile !== null && iconFile !== undefined){
+        try {
+          await axios({
+            url: `/setting/menu/img`,
+            method:'post',
+            baseURL : 'http://localhost:8080',
+            headers: { 'Content-Type': 'multipart/form-data' },
+            data:{
+              iconFile:  iconFile
+            }
+          });
+          console.log(iconFile);
+          console.log('successed sending img...');
+        } catch (error) {
+          console.log('failed sending img...');
+        }
+      } else {
+        console.log('image is null');
+      }
+      
+      */
+
+export function iconListAPI(){
+    return axios({
+      url: `/setting/menu/iconList`,
+      method:'get',
+      baseURL : 'http://localhost:8080',
+    });
 }
